Integrate image safety into LM workflow
